requirements.txt
assemblyai
beautifulsoup4
chromedriver-autoinstaller
google-generativeai
instaloader
python-dotenv
rich
selenium